fix: perspective corrected interpolate

+++ include/utils/buffer.hpp
@@ -7,24 +7,26 @@
 template <typename T>
 class Buffer {
    public:
-    const int width;
-    const int height;
-    const int n_channels;
+    const size_t width;
+    const size_t height;
+    const size_t n_channels;
     T *data;
 
-    Buffer(const int width, const int height, const int n_channels = 1);
-    Buffer(const int width, const int height, const int n_channels,
+    Buffer(const size_t width, const size_t height,
+           const size_t n_channels = 1);
+    Buffer(const size_t width, const size_t height, const size_t n_channels,
            const T init_val);
     ~Buffer();
 
-    inline T &at(const int x, const int y, const int channel = 0) const {
+    inline T &at(const size_t x, const size_t y,
+                 const size_t channel = 0) const {
         return data[(y * width + x) * n_channels + channel];
     }
 };
 
 template <typename T>
-Buffer<T>::Buffer(const int width, const int height, const int n_channels,
-                  const T init_val)
+Buffer<T>::Buffer(const size_t width, const size_t height,
+                  const size_t n_channels, const T init_val)
     : width(width),
       height(height),
       n_channels(n_channels),
@@ -33,7 +35,8 @@ Buffer<T>::Buffer(const int width, const int height, const int n_channels,
 }
 
 template <typename T>
-Buffer<T>::Buffer(const int width, const int height, const int n_channels)
+Buffer<T>::Buffer(const size_t width, const size_t height,
+                  const size_t n_channels)
     : width(width),
       height(height),
       n_channels(n_channels),
